// api/models/AuthRequestOtp200Response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Laravel
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { AuthRequestOtp403ResponseMessage } from './AuthRequestOtp403ResponseMessage';
import {
    AuthRequestOtp403ResponseMessageFromJSON,
    AuthRequestOtp403ResponseMessageFromJSONTyped,
    AuthRequestOtp403ResponseMessageToJSON,
    AuthRequestOtp403ResponseMessageToJSONTyped,
} from './AuthRequestOtp403ResponseMessage';
import type { AuthRequestOtp200ResponseData } from './AuthRequestOtp200ResponseData';
import {
    AuthRequestOtp200ResponseDataFromJSON,
    AuthRequestOtp200ResponseDataFromJSONTyped,
    AuthRequestOtp200ResponseDataToJSON,
    AuthRequestOtp200ResponseDataToJSONTyped,
} from './AuthRequestOtp200ResponseData';

/**
 * 
 * @export
 * @interface AuthRequestOtp200Response
 */
export interface AuthRequestOtp200Response {
    /**
     * 
     * @type {boolean}
     * @memberof AuthRequestOtp200Response
     */
    success: boolean;
    /**
     * 
     * @type {AuthRequestOtp403ResponseMessage}
     * @memberof AuthRequestOtp200Response
     */
    message: AuthRequestOtp403ResponseMessage | null;
    /**
     * 
     * @type {AuthRequestOtp200ResponseStatusEnum}
     * @memberof AuthRequestOtp200Response
     */
    status: AuthRequestOtp200ResponseStatusEnum;
    /**
     * 
     * @type {AuthRequestOtp200ResponseData}
     * @memberof AuthRequestOtp200Response
     */
    data: AuthRequestOtp200ResponseData;
}


/**
 * @export
 */
export const AuthRequestOtp200ResponseStatusEnum = {
    NUMBER_200: 200
} as const;
export type AuthRequestOtp200ResponseStatusEnum = typeof AuthRequestOtp200ResponseStatusEnum[keyof typeof AuthRequestOtp200ResponseStatusEnum];


/**
 * Check if a given object implements the AuthRequestOtp200Response interface.
 */
export function instanceOfAuthRequestOtp200Response(value: object): value is AuthRequestOtp200Response {
    if (!('success' in value) || value['success'] === undefined) return false;
    if (!('message' in value) || value['message'] === undefined) return false;
    if (!('status' in value) || value['status'] === undefined) return false;
    if (!('data' in value) || value['data'] === undefined) return false;
    return true;
}

export function AuthRequestOtp200ResponseFromJSON(json: any): AuthRequestOtp200Response {
    return AuthRequestOtp200ResponseFromJSONTyped(json, false);
}

export function AuthRequestOtp200ResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): AuthRequestOtp200Response {
    if (json == null) {
        return json;
    }
    return {
        
        'success': json['success'],
        'message': AuthRequestOtp403ResponseMessageFromJSON(json['message']),
        'status': json['status'],
        'data': AuthRequestOtp200ResponseDataFromJSON(json['data']),
    };
}

export function AuthRequestOtp200ResponseToJSON(json: any): AuthRequestOtp200Response {
    return AuthRequestOtp200ResponseToJSONTyped(json, false);
}

export function AuthRequestOtp200ResponseToJSONTyped(value?: AuthRequestOtp200Response | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'success': value['success'],
        'message': AuthRequestOtp403ResponseMessageToJSON(value['message']),
        'status': value['status'],
        'data': AuthRequestOtp200ResponseDataToJSON(value['data']),
    };
}