// api/models/AuthVerifyOtp200Response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Laravel
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { AuthRequestOtp403ResponseMessage } from './AuthRequestOtp403ResponseMessage';
import {
    AuthRequestOtp403ResponseMessageFromJSON,
    AuthRequestOtp403ResponseMessageFromJSONTyped,
    AuthRequestOtp403ResponseMessageToJSON,
    AuthRequestOtp403ResponseMessageToJSONTyped,
} from './AuthRequestOtp403ResponseMessage';
import type { AuthVerifyOtp200ResponseData } from './AuthVerifyOtp200ResponseData';
import {
    AuthVerifyOtp200ResponseDataFromJSON,
    AuthVerifyOtp200ResponseDataFromJSONTyped,
    AuthVerifyOtp200ResponseDataToJSON,
    AuthVerifyOtp200ResponseDataToJSONTyped,
} from './AuthVerifyOtp200ResponseData';

/**
 * 
 * @export
 * @interface AuthVerifyOtp200Response
 */
export interface AuthVerifyOtp200Response {
    /**
     * 
     * @type {boolean}
     * @memberof AuthVerifyOtp200Response
     */
    success: boolean;
    /**
     * 
     * @type {AuthRequestOtp403ResponseMessage}
     * @memberof AuthVerifyOtp200Response
     */
    message: AuthRequestOtp403ResponseMessage | null;
    /**
     * 
     * @type {AuthVerifyOtp200ResponseStatusEnum}
     * @memberof AuthVerifyOtp200Response
     */
    status: AuthVerifyOtp200ResponseStatusEnum;
    /**
     * 
     * @type {AuthVerifyOtp200ResponseData}
     * @memberof AuthVerifyOtp200Response
     */
    data: AuthVerifyOtp200ResponseData;
}


/**
 * @export
 */
export const AuthVerifyOtp200ResponseStatusEnum = {
    NUMBER_200: 200
} as const;
export type AuthVerifyOtp200ResponseStatusEnum = typeof AuthVerifyOtp200ResponseStatusEnum[keyof typeof AuthVerifyOtp200ResponseStatusEnum];


/**
 * Check if a given object implements the AuthVerifyOtp200Response interface.
 */
export function instanceOfAuthVerifyOtp200Response(value: object): value is AuthVerifyOtp200Response {
    if (!('success' in value) || value['success'] === undefined) return false;
    if (!('message' in value) || value['message'] === undefined) return false;
    if (!('status' in value) || value['status'] === undefined) return false;
    if (!('data' in value) || value['data'] === undefined) return false;
    return true;
}

export function AuthVerifyOtp200ResponseFromJSON(json: any): AuthVerifyOtp200Response {
    return AuthVerifyOtp200ResponseFromJSONTyped(json, false);
}

export function AuthVerifyOtp200ResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): AuthVerifyOtp200Response {
    if (json == null) {
        return json;
    }
    return {
        
        'success': json['success'],
        'message': AuthRequestOtp403ResponseMessageFromJSON(json['message']),
        'status': json['status'],
        'data': AuthVerifyOtp200ResponseDataFromJSON(json['data']),
    };
}

export function AuthVerifyOtp200ResponseToJSON(json: any): AuthVerifyOtp200Response {
    return AuthVerifyOtp200ResponseToJSONTyped(json, false);
}

export function AuthVerifyOtp200ResponseToJSONTyped(value?: AuthVerifyOtp200Response | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'success': value['success'],
        'message': AuthRequestOtp403ResponseMessageToJSON(value['message']),
        'status': value['status'],
        'data': AuthVerifyOtp200ResponseDataToJSON(value['data']),
    };
}